refactor(assistant): split read tools and decompose cli adapter turn loop
